Move new common JEC functions to PatUtilities.  Fix bug in application of new JEC.
