+++ plugins/TopPairMuonPlusJetsSelectionFilter.cc
@@ -261,7 +261,7 @@ void TopPairMuonPlusJetsSelectionFilter::setupEventContent(edm::Event& iEvent, c
 
 	if ( applyJEC_ ) {
 		corrector_ = JetCorrector::getJetCorrector (jetCorrectionService_, iSetup);
-		jets_ = applyNewJec( jets_, iEvent, iSetup );
+		jets_ = applyNewJec( jets_, corrector_, iEvent, iSetup );
 	}
 
 	// Electrons (for veto)
@@ -448,24 +448,6 @@ void TopPairMuonPlusJetsSelectionFilter::cleanedJets() {
 		}
 		indexInNtuple++;		
 	}
-}
-
-pat::JetCollection TopPairMuonPlusJetsSelectionFilter::applyNewJec( pat::JetCollection jets, edm::Event& iEvent, const edm::EventSetup& iSetup ) {
-	pat::JetCollection newJets;
-
-	for (unsigned index = 0; index < jets.size(); ++index) {
-		const pat::Jet jet = jets.at(index);
-		float JEC = getJECForJet( jet, iEvent, iSetup);
-		pat::Jet newJet( jet );
-		newJet.setP4( jet.p4() * JEC );
-		newJets.push_back( newJet );
-	}
-
-	return newJets;
-}
-
-float TopPairMuonPlusJetsSelectionFilter::getJECForJet(const pat::Jet jet, edm::Event& iEvent, const edm::EventSetup& iSetup ) {
-	return corrector_->correction( jet.correctedJet("Uncorrected"), iEvent, iSetup );
 }
 
 void TopPairMuonPlusJetsSelectionFilter::cleanedBJets() {
